Solution must be an integer

+++ RandomCodes/Summing_the_n_series/Solution.py
@@ -1,7 +1,7 @@
 def sum(n):
     a = 1
     d = 2
-    print( (n/2) * ( 2*a + (n-1)*d ) )
+    print( int((n/2) * ( 2*a + (n-1)*d )) )
 
 T = int(input())
 for i in range(T):
